# openapi/agent-gpt.yaml
openapi: 3.1.0
info:
  title: Ultrasound Ledger Agent API
  version: 1.1.0
  description: >
    Create and edit ultrasound case reports and attach screenshot images.
    Prefer getCaseImageUploadLink for ChatGPT (user opens a browser upload page).
    Base64 attach is legacy and rejects invalid/placeholder payloads.
servers:
  - url: https://ultrasound.margies.app
security:
  - ApiKeyAuth: []
paths:
  /api/agent/cases:
    get:
      operationId: listCases
      summary: Search or list case reports
      parameters:
        - name: exam_date
          in: query
          required: false
          schema:
            type: string
          description: Exam date YYYY-MM-DD
        - name: pet_name
          in: query
          required: false
          schema:
            type: string
        - name: owner_surname
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Matching cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Case'
    post:
      operationId: createCase
      summary: Create a case report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreate'
      responses:
        '201':
          description: Created case (includes upload_url for screenshots)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCreated'
  /api/agent/cases/{id}:
    get:
      operationId: getCase
      summary: Get one case report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
    patch:
      operationId: updateCase
      summary: Edit a case report (not allowed if invoiced)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CasePatch'
      responses:
        '200':
          description: Updated case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
  /api/agent/cases/{id}/upload-link:
    get:
      operationId: getCaseImageUploadLink
      summary: Get a browser upload link for screenshots
      description: >
        PREFERRED for ChatGPT. Returns a short-lived upload_url. Tell the user
        to open that link and select image files. No Base64 conversion needed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Case UUID
      responses:
        '200':
          description: Upload link
          content:
            application/json:
              schema:
                type: object
                properties:
                  case_id:
                    type: string
                  upload_url:
                    type: string
                  expires_in_seconds:
                    type: integer
                  instructions:
                    type: string
                required:
                  - case_id
                  - upload_url
  /api/agent/cases/{id}/images:
    get:
      operationId: listCaseImages
      summary: List images attached to a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image metadata list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseImage'
    post:
      operationId: attachCaseImage
      summary: Attach image via Base64 or file_url (legacy)
      description: >
        Prefer getCaseImageUploadLink. If using this action, provide either a
        real https file_url OR valid data_base64 image bytes. Never invent
        placeholder Base64.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageAttach'
      responses:
        '201':
          description: Attached image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseImage'
  /api/agent/cases/{id}/images/from-url:
    post:
      operationId: attachCaseImageFromUrl
      summary: Attach image by downloading a public HTTPS URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_url
              properties:
                file_url:
                  type: string
                  description: Public HTTPS URL of the image
                filename:
                  type: string
      responses:
        '201':
          description: Attached image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseImage'
  /api/agent/cases/{id}/images/upload:
    post:
      operationId: attachCaseImageFile
      summary: Attach image via multipart file upload
      description: >
        Multipart form field name is "file". ChatGPT Actions often cannot send
        true binary; prefer getCaseImageUploadLink instead.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                filename:
                  type: string
      responses:
        '201':
          description: Attached image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseImage'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Case:
      type: object
      properties:
        id:
          type: string
        exam_date:
          type: string
        owner_surname:
          type: string
        pet_name:
          type: string
        species:
          type: string
        exam_type:
          type: string
        findings_text:
          type: string
        conclusion_text:
          type: string
        image_notes:
          type: string
        standard_fee:
          type: number
        actual_fee:
          type: number
        is_free:
          type: boolean
        free_reason:
          type: string
        billing_note:
          type: string
        invoice_id:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - exam_date
        - owner_surname
        - pet_name
        - exam_type
        - findings_text
        - conclusion_text
    CaseCreated:
      type: object
      properties:
        id:
          type: string
        exam_date:
          type: string
        owner_surname:
          type: string
        pet_name:
          type: string
        species:
          type: string
        exam_type:
          type: string
        findings_text:
          type: string
        conclusion_text:
          type: string
        image_notes:
          type: string
        standard_fee:
          type: number
        actual_fee:
          type: number
        is_free:
          type: boolean
        free_reason:
          type: string
        billing_note:
          type: string
        invoice_id:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        upload_url:
          type: string
          description: Open this link to attach screenshots in a browser
        case_url:
          type: string
      required:
        - id
        - exam_date
        - owner_surname
        - pet_name
        - exam_type
        - findings_text
        - conclusion_text
        - upload_url    CaseCreate:
      type: object
      required:
        - exam_date
        - owner_surname
        - pet_name
        - exam_type
      properties:
        exam_date:
          type: string
        owner_surname:
          type: string
        pet_name:
          type: string
        species:
          type: string
        exam_type:
          type: string
        findings_text:
          type: string
        conclusion_text:
          type: string
        image_notes:
          type: string
        standard_fee:
          type: number
        actual_fee:
          type: number
        is_free:
          type: boolean
        free_reason:
          type: string
        billing_note:
          type: string
    CasePatch:
      type: object
      properties:
        exam_date:
          type: string
        owner_surname:
          type: string
        pet_name:
          type: string
        species:
          type: string
        exam_type:
          type: string
        findings_text:
          type: string
        conclusion_text:
          type: string
        image_notes:
          type: string
        standard_fee:
          type: number
        actual_fee:
          type: number
        is_free:
          type: boolean
        free_reason:
          type: string
        billing_note:
          type: string
    ImageAttach:
      type: object
      properties:
        filename:
          type: string
        content_type:
          type: string
          description: image/jpeg, image/png, image/webp, or image/gif
        data_base64:
          type: string
          description: Real base64 image bytes only — never placeholder text
        file_url:
          type: string
          description: Alternative to data_base64 — public HTTPS image URL
    CaseImage:
      type: object
      properties:
        id:
          type: string
        case_id:
          type: string
        original_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        sort_order:
          type: integer
        created_at:
          type: string
        url:
          type: string
        success:
          type: boolean
      required:
        - id
        - case_id
        - original_name
        - mime_type
        - size_bytes
        - success